Quick fix for roslyn
When entering the buffer, it would go through the whole process of
finding the solutions files again, making it very slow to jump between
buffers. Now we check if the server is started first and simply attach
the server to the buffer.

On windows, vim.fs.root behaves weird and would return nil when finding
the root folder of the sln. A quick fix is to just wrap the sln file in
vim.fs.dirname(sln_file).